+++ src/main/java/heronarts/lx/studio/TEApp.java
@@ -80,6 +80,7 @@ import titanicsend.pattern.justin.*;
 import titanicsend.pattern.look.*;
 import titanicsend.pattern.mike.*;
 import titanicsend.pattern.pixelblaze.*;
+import titanicsend.pattern.sinas.LightBeamsAudioReactivePattern;
 import titanicsend.pattern.tom.*;
 import titanicsend.pattern.util.TargetPixelStamper;
 import titanicsend.pattern.will.PowerDebugger;
@@ -481,6 +482,7 @@ public class TEApp extends LXStudio {
       l.addPattern(TurbulenceLines.class, covPanelPartial, cPalette, up);
       l.addPattern(ShaderEdgesPatternConfig.NeonRipplesEdges.class, covPanelPartial, cPalette, up);
       l.addPattern(ArcEdges.class, covPanelPartial, cPalette, up);
+      l.addPattern(LightBeamsAudioReactivePattern.class, covPanelPartial, cPalette, up);
 
       return l;
     }
