+++ src/main/java/titanicsend/pattern/yoffa/config/ShaderPanelsPatternConfig.java
@@ -3,15 +3,9 @@ package titanicsend.pattern.yoffa.config;
 import heronarts.lx.LX;
 import heronarts.lx.LXCategory;
 import heronarts.lx.parameter.LXParameter;
-import java.util.Arrays;
-import java.util.List;
 import titanicsend.pattern.glengine.ConstructedShaderPattern;
-import titanicsend.pattern.glengine.GLShader;
 import titanicsend.pattern.jon.TEControlTag;
 import titanicsend.pattern.yoffa.framework.TEShaderView;
-import titanicsend.util.GaussianFilter;
-import titanicsend.util.SignalLogger;
-import titanicsend.util.TE;
 
 @SuppressWarnings("unused")
 public class ShaderPanelsPatternConfig {
@@ -34,107 +28,16 @@ public class ShaderPanelsPatternConfig {
   // multiple
   @LXCategory("Native Shaders Panels")
   public static class LightBeamsPattern extends ConstructedShaderPattern {
-    // Internally keeps track of what rotation should be applied to the shader texture.
-    private float degrees = 0;
-
-    // Kept internally to keep track of the time difference applied to the speed of the shader
-    // based on the trebleLevel.
-    private float timeDiff = 0;
-
-    // Gaussian filter used to smoothen out the bassLevel
-    GaussianFilter degreesFilter = new GaussianFilter(10);
-
-    // Gaussian filter used to smoothen out the trebleLevel
-    GaussianFilter timeDiffFilter = new GaussianFilter(20);
-    float levelReactivityControl = 0;
-    float freqReactivityControl = 0;
-
     public LightBeamsPattern(LX lx) {
       super(lx, TEShaderView.ALL_PANELS_INDIVIDUAL);
     }
 
     @Override
     protected void createShader() {
-      controls.setRange(TEControlTag.SPEED, 0.02, -2, 2);
-      controls.setRange(TEControlTag.SPIN, 0.05, -1.0, 1.0);
-      controls.setRange(TEControlTag.LEVELREACTIVITY, 1.2, 0, 2);
-      controls.setRange(TEControlTag.FREQREACTIVITY, 0.01, 0, 1);
+      controls.setRange(TEControlTag.SPEED, 0, -4, 4); // speed
+      controls.setValue(TEControlTag.SPEED, 0.5);
 
-      addShader(
-          "light_beams.fs",
-          new GLShaderFrameSetup() {
-            @Override
-            public void OnFrame(GLShader s) {
-              // Update the pattern local control values based on the UI values.
-              levelReactivityControl =
-                  (float) getControls().getControl(TEControlTag.LEVELREACTIVITY).getValue();
-              freqReactivityControl =
-                  (float) getControls().getControl(TEControlTag.FREQREACTIVITY).getValue();
-
-              // Get the rotation angle from the spin. This is the default value for the
-              // iRotationAngle and we're going to apply a diff on this value based on the
-              // bassLevel.
-              double radians = getRotationAngleFromSpin();
-              int spinControlSign =
-                  Float.compare(
-                      (float) getControls().getLXControl(TEControlTag.SPIN).getValue(), 0.0f);
-              radians += getRotationDiff() * spinControlSign;
-              s.setUniform("iRotationAngle", (float) radians);
-
-              // Similar to the rotation, but for speed instead.
-              double currentTime = getTime();
-              currentTime += getTimeDiff();
-              s.setUniform("iTime", (float) currentTime);
-            }
-          });
-    }
-
-    /***
-     * This function updates the internal degree value based on the audio signals.
-     * @return
-     */
-    private float getRotationDiff() {
-      // Use the GaussianFilter to smppthen the bassLevel signal
-      double filteredBassLevel =
-          degreesFilter.applyGaussianFilter((float) bassLevel * levelReactivityControl);
-      // Amplify the filtered level by 5 so it is more visible.
-      float degreeStep = (float) (filteredBassLevel * 5);
-
-      // Update the internal degree value.
-      degrees -= degreeStep;
-      degrees = normalizeDegree(degrees);
-
-      return degreesToRadians(degrees);
-    }
-
-    /**
-     * Utility function to convert degrees to radians.
-     *
-     * @param degrees
-     * @return degree converted to radian.
-     */
-    private float degreesToRadians(float degrees) {
-      return (float) (degrees * Math.PI / 180.0);
-    }
-
-    /**
-     * Function used to normalize the degree values. This normalization keeps the value between [0,
-     * 360] and is not affecting the visuals. Users can call this to keep the degree values sane and
-     * avoid potential overflow problems.
-     *
-     * @param degree The angel in degrees
-     * @return the degree brought back to 0-360
-     */
-    private float normalizeDegree(double degree) {
-      return (degrees % 360 + 360) % 360;
-    }
-
-    private float getTimeDiff() {
-      float filtered =
-          (float) timeDiffFilter.applyGaussianFilter(trebleLevel * freqReactivityControl);
-      timeDiff += filtered;
-
-      return timeDiff;
+      addShader("light_beams.fs");
     }
   }
